enmacs_controller/ems_simulator/requirements.txt
aiohttp
pyyaml
websockets
